GridCommands: extract search command to a separate file
